Uwe Hermann fixed some warnings.
git-svn-id: svn://svn.berlios.de/openocd/trunk@360 b42882b7-edfa-0310-969c-e2dbd0fdcd60

+++ src/jtag/amt_jtagaccel.c
@@ -60,7 +60,7 @@
 #include "log.h"
 
 /* configuration */
-unsigned long amt_jtagaccel_port;
+u16 amt_jtagaccel_port;
 
 /* interface variables
  */
